+++ library/math/acos.c
@@ -33,7 +33,7 @@ acos(double x) {
         GET_LOW_WORD(lx, x);
         if (((ix - 0x3ff00000) | lx) == 0) {    /* |x|==1 */
             if (hx > 0) return 0.0;        /* acos(1) = 0  */
-            else return pi + 2.0 * pio2_lo;    /* acos(-1)= pi */
+            else return pi + (double) 2.0 * pio2_lo;    /* acos(-1)= pi */
         }
         return (x - x) / (x - x);        /* acos(|x|>1) is NaN */
     }
@@ -45,15 +45,15 @@ acos(double x) {
         r = p / q;
         return pio2_hi - (x - (pio2_lo - x * r));
     } else if (hx < 0) {        /* x < -0.5 */
-        z = (one + x) * 0.5;
+        z = (one + x) * (double) 0.5;
         p = z * (pS0 + z * (pS1 + z * (pS2 + z * (pS3 + z * (pS4 + z * pS5)))));
         q = one + z * (qS1 + z * (qS2 + z * (qS3 + z * qS4)));
         s = sqrt(z);
         r = p / q;
         w = r * s - pio2_lo;
-        return pi - 2.0 * (s + w);
+        return pi - (double) 2.0 * (s + w);
     } else {            /* x > 0.5 */
-        z = (one - x) * 0.5;
+        z = (one - x) * (double) 0.5;
         s = sqrt(z);
         df = s;
         SET_LOW_WORD(df, 0);
@@ -62,6 +62,6 @@ acos(double x) {
         q = one + z * (qS1 + z * (qS2 + z * (qS3 + z * qS4)));
         r = p / q;
         w = r * s + c;
-        return 2.0 * (df + w);
+        return (double) 2.0 * (df + w);
     }
 }
