+++ library/math/asinf.c
@@ -1,17 +1,5 @@
 /*
- * $Id: math_asinf.c,v 1.3 2006-01-08 12:04:23 clib2devs Exp $
- *
-
- *
- *
- * PowerPC math library based in part on work by Sun Microsystems
- * Copyright (C) 1993 by Sun Microsystems, Inc. All rights reserved.
- *
- * Developed at SunPro, a Sun Microsystems, Inc. business.
- * Permission to use, copy, modify, and distribute this
- * software is freely granted, provided that this notice
- * is preserved.
- *
+ * $Id: math_asinf.c,v 1.4 2023-07-16 12:04:23 clib2devs Exp $
  */
 
 #ifndef _MATH_HEADERS_H
@@ -37,12 +25,12 @@ asinf(float x) {
     int32_t hx, ix;
     GET_FLOAT_WORD(hx, x);
     ix = hx & 0x7fffffff;
-    if (ix >= 0x3f800000) {        /* |x| >= 1 */
-        if (ix == 0x3f800000)        /* |x| == 1 */
-            return x * pio2;        /* asin(+-1) = +-pi/2 with inexact */
-        return (x - x) / (x - x);        /* asin(|x|>1) is NaN */
-    } else if (ix < 0x3f000000) {    /* |x|<0.5 */
-        if (ix < 0x39800000) {        /* |x| < 2**-12 */
+    if (ix >= 0x3f800000) {             /* |x| >= 1 */
+        if (ix == 0x3f800000)           /* |x| == 1 */
+            return x * (float) pio2;    /* asin(+-1) = +-pi/2 with inexact */
+        return (x - x) / (x - x);       /* asin(|x|>1) is NaN */
+    } else if (ix < 0x3f000000) {       /* |x|<0.5 */
+        if (ix < 0x39800000) {          /* |x| < 2**-12 */
             if (huge + x > one) return x;/* return x with inexact if x!=0*/
         }
         t = x * x;
@@ -52,12 +40,12 @@ asinf(float x) {
         return x + x * w;
     }
     /* 1> |x|>= 0.5 */
-    w = one - fabsf(x);
+    w = one - (float) fabsf(x);
     t = w * (float) 0.5;
     p = t * (pS0 + t * (pS1 + t * pS2));
     q = one + t * qS1;
     s = sqrt(t);
     w = p / q;
-    t = pio2 - 2.0 * (s + s * w);
+    t = pio2 - (double) 2.0 * (s + s * (double) w);
     if (hx > 0) return t; else return -t;
 }
