+++ library/math/casin.c
@@ -7,11 +7,10 @@
 #endif /* _COMPLEX_HEADERS_H */
 
 double complex
-
 casin(double complex z) {
     double x = creal(z), y = cimag(z);
     double complex
-            w = CMPLX(1.0 - (x - y) * (x + y), -2.0 * x * y),
+            w = CMPLX((double) 1.0 - (x - y) * (x + y), (double) -2.0 * x * y),
             r = clog(CMPLX(-y, x) + csqrt(w));
     return CMPLX(cimag(r), -creal(r));
 }
