+++ library/math/hypot.c
@@ -34,7 +34,7 @@ hypot(double x, double y) {
         if (ha >= 0x7ff00000) {    /* Inf or NaN */
             uint32_t low;
             /* Use original arg order iff result is NaN; quieten sNaNs. */
-            w = fabs(x + 0.0) - fabs(y + 0.0);
+            w = fabs(x + (double) 0.0) - fabs(y + (double) 0.0);
             GET_LOW_WORD(low, a);
             if (((ha & 0xfffff) | low) == 0) w = a;
             GET_LOW_WORD(low, b);
