+++ library/math/log.c
@@ -7,21 +7,20 @@
 #endif /* _MATH_HEADERS_H */
 
 static const double
-	ln2_hi = 6.93147180369123816490e-01, /* 3fe62e42 fee00000 */
-	ln2_lo = 1.90821492927058770002e-10, /* 3dea39ef 35793c76 */
-	two54 = 1.80143985094819840000e+16,	 /* 43500000 00000000 */
-	Lg1 = 6.666666666666735130e-01,		 /* 3FE55555 55555593 */
-	Lg2 = 3.999999999940941908e-01,		 /* 3FD99999 9997FA04 */
-	Lg3 = 2.857142874366239149e-01,		 /* 3FD24924 94229359 */
-	Lg4 = 2.222219843214978396e-01,		 /* 3FCC71C5 1D8E78AF */
-	Lg5 = 1.818357216161805012e-01,		 /* 3FC74664 96CB03DE */
-	Lg6 = 1.531383769920937332e-01,		 /* 3FC39A09 D078C69F */
-	Lg7 = 1.479819860511658591e-01,		 /* 3FC2F112 DF3E5244 */
-	zero = 0.0;
+        ln2_hi = 6.93147180369123816490e-01, /* 3fe62e42 fee00000 */
+        ln2_lo = 1.90821492927058770002e-10, /* 3dea39ef 35793c76 */
+        two54 = 1.80143985094819840000e+16,     /* 43500000 00000000 */
+        Lg1 = 6.666666666666735130e-01,         /* 3FE55555 55555593 */
+        Lg2 = 3.999999999940941908e-01,         /* 3FD99999 9997FA04 */
+        Lg3 = 2.857142874366239149e-01,         /* 3FD24924 94229359 */
+        Lg4 = 2.222219843214978396e-01,         /* 3FCC71C5 1D8E78AF */
+        Lg5 = 1.818357216161805012e-01,         /* 3FC74664 96CB03DE */
+        Lg6 = 1.531383769920937332e-01,         /* 3FC39A09 D078C69F */
+        Lg7 = 1.479819860511658591e-01,         /* 3FC2F112 DF3E5244 */
+        zero = 0.0;
 
 double
-log(double x)
-{
+log(double x) {
     double hfsq, f, s, z, R, w, t1, t2, dk;
     int k, hx, i, j;
     unsigned int lx;
@@ -50,34 +49,31 @@ log(double x)
     SET_HIGH_WORD(x, hx | (i ^ 0x3ff00000)); /* normalize x or x/2 */
 
     k += (i >> 20);
-    f = x - 1.0;
+    f = x - (double) 1.0;
 
     if ((0x000fffff & (2 + hx)) < 3) /* |f| < 2**-20 */
     {
-        if (f == zero)
-        {
+        if (f == zero) {
             if (k == 0)
                 return zero;
-            else
-            {
-                dk = (double)k;
+            else {
+                dk = (double) k;
                 return dk * ln2_hi + dk * ln2_lo;
             }
         }
 
-        R = f * f * (0.5 - 0.33333333333333333 * f);
+        R = f * f * ((double) 0.5 - (double) 0.33333333333333333 * f);
 
         if (k == 0)
             return f - R;
-        else
-        {
-            dk = (double)k;
+        else {
+            dk = (double) k;
             return dk * ln2_hi - ((R - dk * ln2_lo) - f);
         }
     }
 
-    s = f / (2.0 + f);
-    dk = (double)k;
+    s = f / ((double) 2.0 + f);
+    dk = (double) k;
     z = s * s;
     i = hx - 0x6147a;
     w = z * z;
@@ -87,16 +83,13 @@ log(double x)
     i |= j;
     R = t2 + t1;
 
-    if (i > 0)
-    {
-        hfsq = 0.5 * f * f;
+    if (i > 0) {
+        hfsq = (double) 0.5 * f * f;
         if (k == 0)
             return f - (hfsq - s * (hfsq + R));
         else
             return dk * ln2_hi - ((hfsq - (s * (hfsq + R) + dk * ln2_lo)) - f);
-    }
-    else
-    {
+    } else {
         if (k == 0)
             return f - s * (f - R);
         else
